factor out core subpackage
