+++ setup.cfg
@@ -41,8 +41,7 @@ include_package_data = True
 python_requires = >=3.12.0
 packages =
     voice_commander
-    voice_commander.core
-    voice_commander.core.schema
+    voice_commander.schema
 
 install_requires =
     pydantic
