+++ voice_commander/conditions.py
